[MINOR] Cosmetic changes for names and log msgs
