+++ hudi-flink-datasource/hudi-flink/src/main/java/org/apache/hudi/source/rebalance/selector/StreamReadAppendKeySelector.java
@@ -16,16 +16,16 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.source.filedistribution.selector;
+package org.apache.hudi.source.rebalance.selector;
 
 import org.apache.hudi.table.format.mor.MergeOnReadInputSplit;
 
 import org.apache.flink.api.java.functions.KeySelector;
 
-public class StreamReadBucketIndexKeySelector implements KeySelector<MergeOnReadInputSplit, String> {
+public class StreamReadAppendKeySelector implements KeySelector<MergeOnReadInputSplit, Integer> {
 
   @Override
-  public String getKey(MergeOnReadInputSplit mergeOnReadInputSplit) throws Exception {
-    return mergeOnReadInputSplit.getFileId();
+  public Integer getKey(MergeOnReadInputSplit mergeOnReadInputSplit) throws Exception {
+    return mergeOnReadInputSplit.getSplitNumber();
   }
 }
