+++ hudi-flink-datasource/hudi-flink/src/main/java/org/apache/hudi/table/HoodieTableSource.java
@@ -46,10 +46,10 @@ import org.apache.hudi.source.FileIndex;
 import org.apache.hudi.source.IncrementalInputSplits;
 import org.apache.hudi.source.StreamReadMonitoringFunction;
 import org.apache.hudi.source.StreamReadOperator;
-import org.apache.hudi.source.filedistribution.partitioner.StreamReadAppendPartitioner;
-import org.apache.hudi.source.filedistribution.partitioner.StreamReadBucketIndexPartitioner;
-import org.apache.hudi.source.filedistribution.selector.StreamReadAppendKeySelector;
-import org.apache.hudi.source.filedistribution.selector.StreamReadBucketIndexKeySelector;
+import org.apache.hudi.source.rebalance.partitioner.StreamReadAppendPartitioner;
+import org.apache.hudi.source.rebalance.partitioner.StreamReadBucketIndexPartitioner;
+import org.apache.hudi.source.rebalance.selector.StreamReadAppendKeySelector;
+import org.apache.hudi.source.rebalance.selector.StreamReadBucketIndexKeySelector;
 import org.apache.hudi.source.prune.DataPruner;
 import org.apache.hudi.source.prune.PartitionPruners;
 import org.apache.hudi.source.prune.PrimaryKeyPruners;
